+++ NEWS.md
@@ -110,6 +110,7 @@ The changes are reflected in the installation section of the handbook.
     - various optimizations to load (cached) EIC data
 - Fixed: `unset()` for `featureGroupsScreeningSet` resulted in loss of group quality scores and internal standard assignments
 - Several optimizations for `annotatedPeakLists()`, especially with sets workflows.
+- `plotInt()`: `plotArgs` and `linesArgs` to pass additional arguments to `plot()`/`lines()`. The latter replaces the dots argument.
 
 
 # patRoon 2.1